+++ c/MonteCarlo.h
@@ -1,2 +1,2 @@
-double MonteCarlo_integrate(int Num_samples);
-double MonteCarlo_num_flops(int Num_samples);
+double MonteCarlo_integrate(unsigned long Num_samples);
+double MonteCarlo_num_flops(unsigned long Num_samples);
